+++ Tests/VpnHood.Test/TestConstants.cs
@@ -1,6 +1,5 @@
 ﻿using System.Diagnostics;
 using System.Net;
-using VpnHood.Core.Toolkit.Converters;
 using VpnHood.Core.Toolkit.Utils;
 
 namespace VpnHood.Test;
@@ -34,6 +33,6 @@ public class TestConstants
     public static IPAddress PingV6Address1 => IPAddress.Parse("2001:4860:4866::2200");
     public static Uri InvalidUri => new("https://DBBC5764-D452-468F-8301-4B315507318F.zz");
     public static IPAddress InvalidIp => IPAddress.Parse("198.51.100.1");
-    public static IPEndPoint InvalidEp => IPEndPointConverter.Parse("198.51.100.2:9999");
+    public static IPEndPoint InvalidEp => IPEndPoint.Parse("198.51.100.2:9999");
     public static IPAddress BlockedIp => IPAddress.Parse("198.18.255.3");
 }