+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Upcoming
 * Fix: Internal crash in logger when closing the service
+* Fix: VpnService is unreachable
+* Fix: Android: Always-on VPN
+* Feature: Custom Server Address for each access token
 
 # v7.3.734
 * Fix: Drop brodcast packets